requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
codecov==2.1.13
coverage==7.3.0
idna==3.4
iniconfig==2.0.0
lxml==4.9.3
multidict==6.0.4
packaging==23.1
pluggy==1.3.0
pytest==7.4.0
pytest-cov==4.1.0
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
urllib3==2.0.4
vcrpy==5.1.0
wrapt==1.15.0
yarl==1.9.2
